refactor: environment_utils/habitat_utils.py -> simulators/habitat/environment_utils.py
